book の登録削除を admin に移動 #11
